Add generic popover component and action
