+++ lxl-web/src/lib/actions/popover.ts
@@ -1,30 +1,29 @@
 import type { Action } from 'svelte/action';
-import type { LocaleCode } from '$lib/i18n/locales';
-import ResourcePopover from './ResourcePopover.svelte';
-import type { ResourceData } from '$lib/types/resourceData';
+import Popover from '$lib/components/Popover.svelte';
 
 /**
- * Svelte action used for showing decorated resource data inside a popover by providing the resource id.
- * The action takes care of the fetching of the required resource data.
+ * Svelte action used for showing generic content inside a popover. This action will be much more flexible with Svelte 5 when
+ * we will be able to set slots when creating component instances (see: https://github.com/sveltejs/svelte/issues/2588)
  *
  * Tests to do:
- * - [] Doesn't attach popover if fetching of resource data fails
+ * - [] Attaches popover when user hovers over trigger node (after delay)
+ * - [] Removes popover when user stops hovering over trigger node (after delay)
+ * - [] Removes popover when user blurs trigger node
+ * - [] Keeps popover open if user hovers over popover (when leaving trigger node)
+ * - [] Removes popover if user stops hovering over popover (without entering trigger node)
+ * - [] Keeps popover if user has focused child element
+ * - [] Closes popover immediately when the URL changes
  */
 
-export const resourcePopover: Action<
-	HTMLElement,
-	{ id?: string; data?: ResourceData[]; lang: LocaleCode }
-> = (node: HTMLElement, options) => {
-	const FETCH_DELAY = 250;
+export const popover: Action<HTMLElement, { title: string }> = (node: HTMLElement, { title }) => {
 	const ATTACH_DELAY = 500;
 	const REMOVE_DELAY = 200;
 	const container = document.getElementById('floating-elements-container') || document.body; // See https://atfzl.com/articles/don-t-attach-tooltips-to-document-body
 
 	let attached = false;
-	let floatingElement: ResourcePopover | null = null;
+	let floatingElement: Popover | null = null;
 	let cancelAttach: (reason?: unknown) => void;
 	let cancelRemove: (reason?: unknown) => void;
-	let cancelFetch: (reason?: unknown) => void;
 
 	node.addEventListener('mouseover', attachPopover);
 	node.addEventListener('mouseout', removePopover);
@@ -35,20 +34,17 @@ export const resourcePopover: Action<
 		try {
 			cancelAttach?.(); // cancel earlier promises to ensure popovers doesn't appear after navigating
 			cancelRemove?.();
-			cancelFetch?.();
-			if ((options.id || options.data) && !attached) {
-				const [decoratedData] = await Promise.all([
-					getDecoratedData(),
-					new Promise((resolve, reject) => {
-						cancelAttach = reject; // allows promise rejection from outside the promise constructor scope
-						setTimeout(resolve, ATTACH_DELAY);
-					})
-				]);
-				floatingElement = new ResourcePopover({
+			if (!attached) {
+				await new Promise((resolve, reject) => {
+					cancelAttach = reject; // allows promise rejection from outside the promise constructor scope
+					setTimeout(resolve, ATTACH_DELAY);
+				});
+
+				floatingElement = new Popover({
 					target: container,
 					props: {
 						referenceElement: node,
-						data: decoratedData,
+						data: title, // data should be removed when we can set slots directly using Svelte 5 (see: https://github.com/sveltejs/svelte/issues/2588)
 						onMouseOver: startFloatingElementInteraction,
 						onFocus: startFloatingElementInteraction,
 						onMouseLeave: endFloatingElementInteraction,
@@ -64,7 +60,6 @@ export const resourcePopover: Action<
 	async function removePopover() {
 		try {
 			cancelAttach?.();
-			cancelFetch?.();
 			await new Promise((resolve, reject) => {
 				cancelRemove = reject;
 				setTimeout(resolve, REMOVE_DELAY);
@@ -76,7 +71,6 @@ export const resourcePopover: Action<
 
 	function destroyPopover() {
 		cancelAttach?.();
-		cancelFetch?.();
 		cancelRemove?.();
 		floatingElement?.$destroy();
 		attached = false;
@@ -90,25 +84,6 @@ export const resourcePopover: Action<
 		removePopover();
 	}
 
-	async function getDecoratedData() {
-		await new Promise((resolve, reject) => {
-			cancelFetch = reject;
-			setTimeout(resolve, FETCH_DELAY);
-		});
-		try {
-			if (options.data) {
-				return options.data;
-			}
-			if (options.id) {
-				const resourceRes = await fetch(`/api/${options.lang}/${options.id.split('/').pop()}`);
-				return await resourceRes.json();
-			}
-		} catch (error) {
-			console.error(error);
-			cancelAttach?.();
-		}
-	}
-
 	return {
 		destroy() {
 			destroyPopover();
@@ -120,4 +95,4 @@ export const resourcePopover: Action<
 	};
 };
 
-export default resourcePopover;
+export default popover;
